Move chamber_of_deputies models related samples from core tests to chamber tests
