+++ jarbas/api/tests/test_applicant_view.py
@@ -5,7 +5,7 @@ from django.shortcuts import resolve_url
 from django.test import TestCase
 
 from jarbas.chamber_of_deputies.models import Reimbursement
-from jarbas.core.tests import sample_reimbursement_data
+from jarbas.chamber_of_deputies.tests import sample_reimbursement_data
 
 
 class TestApplicant(TestCase):
