+++ jarbas/core/tests/test_tweets_command.py
@@ -7,7 +7,7 @@ from mixer.backend.django import mixer
 
 from jarbas.chamber_of_deputies.models import Reimbursement, Tweet
 from jarbas.core.management.commands.tweets import Command
-from jarbas.core.tests import random_tweet_status
+from jarbas.chamber_of_deputies.tests import random_tweet_status
 
 
 KEYS = (
